Reorganize the module
