+++ src/arg_parsers/identify_format/mod.rs
@@ -1,0 +1,4 @@
+mod identify_format;
+pub use identify_format::*;
+
+mod parser;
